requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   starlette
    #   watchfiles
black==26.1.0
    # via -r requirements.in
build==1.4.0
    # via pip-tools
cfgv==3.5.0
    # via pre-commit
click==8.3.1
    # via
    #   black
    #   pip-tools
    #   uvicorn
coverage[toml]==7.13.4
    # via pytest-cov
distlib==0.4.0
    # via virtualenv
fastapi==0.135.1
    # via -r requirements.in
filelock==3.25.0
    # via
    #   python-discovery
    #   virtualenv
flake8==7.3.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
identify==2.6.17
    # via pre-commit
idna==3.11
    # via anyio
iniconfig==2.3.0
    # via pytest
isort==8.0.1
    # via -r requirements.in
librt==0.8.1
    # via mypy
mccabe==0.7.0
    # via flake8
mypy==1.19.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.10.0
    # via pre-commit
packaging==26.0
    # via
    #   black
    #   build
    #   pytest
    #   wheel
pathspec==1.0.4
    # via
    #   black
    #   mypy
pip-tools==7.5.3
    # via -r requirements.in
platformdirs==4.9.2
    # via
    #   black
    #   python-discovery
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.5.1
    # via -r requirements.in
pycodestyle==2.14.0
    # via flake8
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==7.0.0
    # via -r requirements.in
python-discovery==1.1.0
    # via virtualenv
python-dotenv==1.2.2
    # via uvicorn
pytokens==0.4.1
    # via black
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   pre-commit
    #   uvicorn
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn[standard]==0.41.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
virtualenv==21.1.0
    # via pre-commit
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
